+++ UserDev/ClusterStudy/CMAlgoOutOfConeSeparate.hh
@@ -57,7 +57,7 @@ namespace cluster {
     void SetMinLength(float len) { _MinLen = len; }
 
     /// SetMinimum number of hits for small cluster
-    void SetMinHits(int n) { _minHits = n; }
+    void SetMinHits(size_t n) { _minHits = n; }
 
     /// Function to reset the algorithm instance ... maybe implemented via child class
     virtual void Reset(){}
@@ -72,7 +72,7 @@ namespace cluster {
     float _MaxAngleFar;
     float _MinLen;
     float _FallOff;
-    int _minHits;
+    size_t _minHits;
 
   };
 }
