+++ UserDev/ClusterStudy/CMAlgoPolyOverlap.hh
@@ -44,7 +44,7 @@ namespace cluster {
     void SetDebug(bool debug) { _debug = debug; }
 
     //both clusters must have > this # of hits to be considered for merging
-    void SetMinNumHits(int nhits) { _min_hits = nhits; }
+    void SetMinNumHits(size_t nhits) { _min_hits = nhits; }
 
     /// Method to re-configure the instance
     void reconfigure();
@@ -52,7 +52,7 @@ namespace cluster {
   private:
     
     bool _debug;
-    int _min_hits;
+    size_t _min_hits;
   };
 }
