+++ UserDev/ClusterStudy/CMAlgoStartNearEnd.cc
@@ -34,8 +34,8 @@ namespace cluster {
     double angle_1 = cluster1.GetParams().opening_angle;
     double angle_2 = cluster2.GetParams().opening_angle;
 
-    int hits_1 = cluster1.GetHitVector().size();
-    int hits_2 = cluster2.GetHitVector().size();
+    size_t hits_1 = cluster1.GetHitVector().size();
+    size_t hits_2 = cluster2.GetHitVector().size();
 
 
     if ( (angle_1 < _maxopeningangle) and (hits_1 > _MinHits) and
